+++ resources/assets/scripts/views/ordenes/productos.js
@@ -90,27 +90,47 @@ app || (app = {});
                 model = this.collection.get(resource),
                 _this = this;
 
-            if ( model instanceof Backbone.Model ) {
-                model.destroy({
-                    success : function(model, resp) {
-                        if(!_.isUndefined(resp.success)) {
-                            window.Misc.removeSpinner( _this.parameters.wrapper );
+            // function confirm delete item
+            this.confirmDelete(model);
+        },
 
-                            if( !resp.success ) {
-                                alertify.error(resp.errors);
-                                return;
-                            }
+        /**
+        * modal confirm delete area
+        */
+        confirmDelete: function( model ) {
+            var _this = this;
 
-                            model.view.remove();
-                            _this.collection.remove(model);
+            var cancelConfirm = new window.app.ConfirmWindow({
+                parameters: {
+                    dataFilter: { producto_nombre: model.get('productop_nombre'), producto_id: model.get('id')},
+                    template: _.template( ($('#ordenp-productop-delete-confirm-tpl').html() || '') ),
+                    titleConfirm: 'Eliminar producto',
+                    onConfirm: function () {
+                        if ( model instanceof Backbone.Model ) {
+                            model.destroy({
+                                success : function(model, resp) {
+                                    if(!_.isUndefined(resp.success)) {
+                                        window.Misc.removeSpinner( _this.parameters.wrapper );
 
-                            // Update total
-                            _this.totalize();
+                                        if( !resp.success ) {
+                                            alertify.error(resp.errors);
+                                            return;
+                                        }
+
+                                        model.view.remove();
+                                        _this.collection.remove(model);
+
+                                        // Update total
+                                        _this.totalize();
+                                    }
+                                }
+                            });
                         }
                     }
-                });
+                }
+            });
 
-            }
+            cancelConfirm.render();
         },
 
         /**
